requirements.txt
PyQt6
torch
torchvision
torchaudio
easyocr
transformers
sentencepiece
Pillow>=9.1.0
numpy==1.24.3
matplotlib>=3.9.1
opencv-python-headless>=4.8.0
imageio>=2.35.1
scikit-image>=0.24.0
mediapipe>=0.10.14